Add rust_decimal

+++ matching-engine/src/structs/order.rs
@@ -1,10 +1,11 @@
+use rust_decimal::Decimal;
 use schema::generated::order_requests::OrderSide;
 
 pub struct LimitOrder {
     pub order_id: u64,
     pub user_id: u64,
-    pub price: u64,
-    pub initial_size: u64,
-    pub remaining_size: u64,
+    pub price: Decimal,
+    pub initial_size: Decimal,
+    pub remaining_size: Decimal,
     pub side: OrderSide,
 }
